+++ app/app.css
@@ -5,6 +5,7 @@ body {
   background-color: #232C39;
   background-image: linear-gradient(45deg, rgba(0, 216, 255, .5) 10%, rgba(0, 1, 127, .7));
   font-family: Arial, Helvetica, Helvetica Neue;
+  overflow-y: hidden;
 }
 
 h2 {
