+++ app/containers/App.production.js
@@ -1,0 +1,16 @@
+import React, { Component, PropTypes} from 'react';
+
+export default class App extends Component {
+  static propTypes = {
+    children: PropTypes.element.isRequired
+  };
+
+  render() {
+    return (
+      <div>
+        {this.props.children}
+      </div>
+    );
+  }
+}
+
