+++ app/store/configureStore.js
@@ -1,21 +1,5 @@
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import reducer from '../reducers';
-
-const createStoreWithMiddleware = applyMiddleware(
-  thunk
-)(createStore);
-
-export default function configureStore(initialState) {
-  const store = createStoreWithMiddleware(reducer, initialState);
-
-  if (module.hot) {
-    // Enable Webpack hot module replacement for reducers
-    module.hot.accept('../reducers', () => {
-      const nextReducer = require('../reducers');
-      store.replaceReducer(nextReducer);
-    });
-  }
-
-  return store;
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./configureStore.production');
+} else {
+  module.exports = require('./configureStore.development');
 }
